Start search for reports at the reports header
Unfortunately the reports aren't contained in a useful 'div' because the
start of the div previously also contained the "notable catches"
section. I *think* I've probably not noticed this previously because
using a dictionary for the reports overwrote the prior values for the
spot (e.g. `spot['foo'] = "someone caught a big fish"`, then later
`spot['foo'] = "the actual report for the same spot"`).

Now should just skip to the proper section by looking for the header for
"Statewide Waters" -- hopefully that shouldn't change anytime soon.